Code cleanup.

+++ Sources/Uberball.Game.Logic.Entities/Decoration.cs
@@ -1,7 +1,10 @@
 ﻿
 namespace Uberball.Game.Logic.Entities {
 	public class Decoration {
-		public double X { get; set; }
-		public double Y { get; set; }
+		/// <summary>Gets or sets X coordinate.</summary>
+		public float X { get; set; }
+
+		/// <summary>Gets or sets Y coordinate.</summary>
+		public float Y { get; set; }
 	}
 }
