+++ Sources/Uberball.Game.NetworkProtocol/GroundSerializer.cs
@@ -5,7 +5,11 @@ namespace Uberball.Game.NetworkProtocol {
 	using Khrussk.NetworkRealm.Protocol;
 	using Logic.Entities;
 
+	/// <summary>Ground serializer.</summary>
 	public class GroundSerializer : IEntitySerializer<Ground> {
+		/// <summary>Deserializes ground from stream.</summary>
+		/// <param name="reader">Reader.</param>
+		/// <param name="entity">Entity.</param>
 		public void Deserialize(BinaryReader reader, ref Ground entity) {
 			entity = entity ?? new Ground();
 			var cnt = reader.ReadByte();
@@ -16,12 +20,15 @@ namespace Uberball.Game.NetworkProtocol {
 			}
 		}
 
+		/// <summary>Serializes entity into stream.</summary>
+		/// <param name="writer">Writer.</param>
+		/// <param name="entity">Entity.</param>
 		public void Serialize(BinaryWriter writer, Ground entity) {
 			writer.Write((byte)entity.Points.Count);
 			writer.Write((byte)entity.Type);
 			foreach (var point in entity.Points) {
-				writer.Write((Single)point.X);
-				writer.Write((Single)point.Y);
+				writer.Write(point.X);
+				writer.Write(point.Y);
 			}
 		}
 	}
